Move like post to post controller
